+++ server/model/examine.js
@@ -2,7 +2,7 @@
 
 module.exports = (sequelize, DataTypes) => {
     return sequelize.define('Examine', {
-        userID: {type: DataTypes.INTEGER(8).ZEROFILL, allowNull: false},
+        userID: {type: DataTypes.INTEGER(11).ZEROFILL, allowNull: false},
         examID: {type: DataTypes.INTEGER(8), primaryKey: true, allowNull: false},
         result: {type: DataTypes.JSON, allowNull: false},
         score: {type: DataTypes.FLOAT, allowNull: false}
