+++ server/model/project-scorerecord.js
@@ -6,12 +6,12 @@ module.exports = function(sequelize,DataTypes){
             allowNull:false,
         },
         userID:{
-            type:DataTypes.INTEGER(8).ZEROFILL,
+            type:DataTypes.INTEGER(11).ZEROFILL,
             primaryKey:true,
             allowNull:false,
         },
         scorerID:{
-            type:DataTypes.INTEGER(8).ZEROFILL,
+            type:DataTypes.INTEGER(11).ZEROFILL,
             primaryKey:true,
             allowNull:false,
         },
