+++ server/model/user-information.js
@@ -1,7 +1,7 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
     return sequelize.define('UserInformation', {
-        userID: {type: DataTypes.INTEGER(8).ZEROFILL, primaryKey: true, allowNull: false},
+        userID: {type: DataTypes.INTEGER(11).ZEROFILL, primaryKey: true, allowNull: false},
         birthday: {type: DataTypes.DATE, allowNull: true},
         sex: {type: DataTypes.STRING(1), allowNull: true},
         nickname: {type: DataTypes.STRING(45), allowNull: true},
